removed build warnings so pod can be pushed to spec repository

+++ GTViewController/Classes/Custom Classes/UIRoundedView.m
@@ -113,7 +113,7 @@ extern NSString * const kTitleMode_peek;
         NSTimeInterval currentTimeStamp = [touch timestamp];
 
 		//calculate changes in direction and time since last call to touchesMoved
-        double deltaX                   = fabsf(currentTouchPosition.x - self.lastTouchPosition.x);
+        double deltaX                   = fabs(currentTouchPosition.x - self.lastTouchPosition.x);
         double deltaT                   = currentTimeStamp - self.lastTimeStamp;
 
 		//add the current change in distance to the total distance
